+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Update unopened order "Delete piece" warning message. Refs UIOR-905.
 * Validate All POLs on order that is set to export via EDIFACT have same account number. Refs UIOR-903.
 * Adding tenant's timezone in UI. Refs UIOR-910.
+* Settings > Orders > change focus. Refs UIOR-897.
 
 ## [3.0.3](https://github.com/folio-org/ui-orders/tree/v3.0.3) (2021-12-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v3.0.2...v3.0.3)